JavaDocの修正とtestの修正
