+++ src/test/java/com/u1pena/estimateapi/estimate/controller/EstimateControllerTest.java
@@ -25,8 +25,8 @@ import com.u1pena.estimateapi.estimate.entity.EstimateBase;
 import com.u1pena.estimateapi.estimate.exception.EstimateException.EstimateBaseNotFoundException;
 import com.u1pena.estimateapi.estimate.exception.EstimateException.EstimateProductNotFoundException;
 import com.u1pena.estimateapi.estimate.exception.EstimateException.NoMatchMaintenanceGuideException;
+import com.u1pena.estimateapi.estimate.helper.EstimateTestHelper;
 import com.u1pena.estimateapi.estimate.service.EstimateService;
-import com.u1pena.estimateapi.master.helper.EstimateTestHelper;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
