+++ src/test/java/com/u1pena/estimateapi/estimate/repository/EstimateRepositoryTest.java
@@ -8,9 +8,9 @@ import com.u1pena.estimateapi.estimate.dto.EstimateProductJoinResult;
 import com.u1pena.estimateapi.estimate.dto.EstimateSummaryResult;
 import com.u1pena.estimateapi.estimate.entity.EstimateBase;
 import com.u1pena.estimateapi.estimate.entity.EstimateProduct;
+import com.u1pena.estimateapi.estimate.helper.EstimateTestHelper;
 import com.u1pena.estimateapi.master.entity.GuideProductPermission;
 import com.u1pena.estimateapi.master.entity.MaintenanceGuide;
-import com.u1pena.estimateapi.master.helper.EstimateTestHelper;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
